Add internet proxy server system to relay requests from outside the intranet
A proxy server on the internet can be installed (server/proxy/internet)
that caches all the requests to it, but doesn't decrypt them or
anyhting. A second element of the proxy server is on the intranet (server/proxy/intranet) and it
periodically sends requests to the internet proxy and forwards the
requests to their destinations. That way, requests can be stored
temporarily and batch-processed later on. It also doesn't require any
port forwarding etc. that way.
However, delay may be quite high.

In case this method doesn't prove reliable, the alternative is to make
an SSH tunnel into the intranet for the relevant ports.