API: Automatically re-connect to MongoDB if connection fails, improve MongoDB error reporting, add settings system for per-server setting
Mongoose is horrible: Instead of failing when MongoDB becomes unavailable and reporting the issues to the program, it doesn't do anything and just silently
accepts that the MongoDB server has gone missing. It doesn't even reconnect once the server comes back online. Therefore, the API server uses a custom
timeout that automatically triggers a connection reset if the server doesn't answer in a specific period of time.

However, the issue of transactions that get buffered in mongoose and only then all performed when the server comes back online still remains. So, if the
database server outage takes more than just a couple of seconds, it is propably a good idea to restart all API servers. I will try to set a per-query timeout
next. If I succeed, this will become unneccesary.