API: Fix function counting the number of decimals for transactions
Since JavaScript starts to use scientific notation in the .toString()
function at a certain number of decimals, the previous algorithm broke
at that point. Due to rounding issues, this allowed a malicious attacker
to send thousands of transactions that each deduct 0.00001 HGC from the
target's account which are transferred to the tax income account and
thus stealing money. However, the attacker would not be able to put the
money on his own account.

Thanks to user @mein-account for reporting this issue, the user received
a bug bounty.