API: Properly escape strings provided by user before creating a new RegExp
This should mitigate DoS attacks on the Database Server since malicious regular expressions
that hang the server cannot be created by users (ReDoS). Previously, regular expressions with
long execution times could be sent to the server in order to slow it down.

Great thanks to user @mkuron for reporting the issue.